third_party/WebKit/LayoutTests/http/tests/w3c/webperf/submission/Google/HighResolutionTime/worker-dedicated-basic-expected.txt

Description

This test validates that performance.now() exists in dedicated workers and reports reasonable times.


PASS workerStart not defined on the Worker object 
PASS self.performance is defined 
PASS self.performance.now is defined 
PASS Time in the worker should be after the startTime in the main document 
PASS Time in the worker should be before the current time in the main document 
